Use PATH search instead of variables for executables

+++ agape_annot.sh
@@ -25,33 +25,42 @@ if [ ! -e "$REF_FASTA" ]
 then
   echo "No reference sequence data: annotation based on homology skipped"
 else
-	if [ ! -e "$out_dir"/"$out_name".scf.fasta ]
-	then
-		ln -s $seq $out_dir/$out_name.scf.fasta
-	fi
- 	$SCRIPTS/homology_annot.sh $out_dir $out_name $SCRIPTS # results place in $out_dir/annot/$out_name.codex
+  if [ ! -e "$out_dir"/"$out_name".scf.fasta ]
+  then
+    ln -s $seq $out_dir/$out_name.scf.fasta
+  fi
+  if [ ! -e "$out_dir"/annot/"$out_name".codex ]; then
+    $SCRIPTS/homology_annot.sh $out_dir $out_name $SCRIPTS # results place in $out_dir/annot/$out_name.codex
+  fi
 fi
 
 snap_dir=$out_dir/snap_files
-mkdir -p $snap_dir
-
-ln -s $REF_FASTA $snap_dir/$REF_NAME.dna
-$SCRIPTS/prep_maker.sh $snap_dir $SCRIPTS
+if [ ! -d $snap_dir ]; then
+  mkdir -p $snap_dir;
+  ln -s $REF_FASTA $snap_dir/$REF_NAME.dna
+  $SCRIPTS/prep_maker.sh $snap_dir $SCRIPTS
+fi
 
 maker_dir=$out_dir/maker
-rm -rf $out_dir/maker
-mkdir -p $maker_dir
-
-$SCRIPTS/run_maker.sh $out_dir $out_name $maker_dir $snap_dir $SCRIPTS # results in $maker_dir/genes.gff
+if [ ! -e "$maker_dir"/genes.gff ]; then
+  rm -rf $out_dir/maker
+  mkdir -p $maker_dir
+  $SCRIPTS/run_maker.sh $out_dir $out_name $maker_dir $snap_dir $SCRIPTS # results in $maker_dir/genes.gff
+fi
 
 comb_annot=$out_dir/comb_annot
-rm -rf $comb_annot
-mkdir -p $comb_annot
+# rm -rf $comb_annot
+if [ ! -d $comb_annot ]; then
+  mkdir -p $comb_annot
+fi
 
 cd $comb_annot
-$SCRIPTS/run_comb_annot.sh $comb_annot $BLAST $out_name $maker_dir $out_dir $snap_dir $SCRIPTS
+BLAST='unused'  # Unused
+if [ ! -e $comb_annot/gff/$out_name.genes.gff ]; then
+  $SCRIPTS/run_comb_annot.sh $comb_annot $BLAST $out_name $maker_dir $out_dir $snap_dir $SCRIPTS
+fi
 
-$SCRIPTS/final_annot.sh $out_dir/maker/seq.fasta $comb_annot/gff/$out_name.genes.gff $comb_annot $out_name $SCRIPTS  
+$SCRIPTS/final_annot.sh $out_dir/maker/seq.fasta $comb_annot/gff/$out_name.genes.gff $comb_annot $out_name $SCRIPTS
 
 # The results are $comb_annot/$out_name.gff
 # BLASTX output file is $comb_annot/blast_out/$out_name.blastx.out
